api/requirements.txt
pandas
numpy
scikit-learn
transformers
torch
sentencepiece
requests
fastapi
pydantic
geocoder==1.38.1
python-dotenv==1.0.1
httpx
groq
pytest
pytest-asyncio==0.23.5
uvicorn
SQLAlchemy==2.0.35